Refactor fs2db migration into mlflow.store.fs2db package with CLI command
Move migration logic from fs2db/src/ into mlflow/store/fs2db/ as a proper
package and add `mlflow migrate-filestore` CLI command. Split the monolithic
migrate.py into focused modules, use pathlib throughout, and remove redundant
module docstrings.
