+++ mlflow/store/fs2db/_tracking.py
@@ -1,0 +1,610 @@
+import json
+import math
+import uuid
+from pathlib import Path
+
+from sqlalchemy.orm import Session
+
+from mlflow.entities import RunStatus
+from mlflow.store.fs2db._helpers import (
+    META_YAML,
+    bump,
+    list_experiment_ids,
+    list_files,
+    list_subdirs,
+    read_metric_lines,
+    read_tag_files,
+    safe_read_yaml,
+)
+from mlflow.store.tracking.dbmodels.models import (
+    SqlAssessments,
+    SqlDataset,
+    SqlExperiment,
+    SqlExperimentTag,
+    SqlInput,
+    SqlInputTag,
+    SqlLatestMetric,
+    SqlLoggedModel,
+    SqlLoggedModelMetric,
+    SqlLoggedModelParam,
+    SqlLoggedModelTag,
+    SqlMetric,
+    SqlParam,
+    SqlRun,
+    SqlTag,
+    SqlTraceInfo,
+    SqlTraceMetadata,
+    SqlTraceTag,
+)
+
+# ── Phase 1: Experiments + Tags ──────────────────────────────────────────────
+
+
+def migrate_experiments(session: Session, mlruns: Path) -> None:
+    for exp_id in list_experiment_ids(mlruns):
+        _migrate_one_experiment(session, mlruns / exp_id, exp_id)
+
+    trash_dir = mlruns / ".trash"
+    for exp_id in list_experiment_ids(trash_dir):
+        _migrate_one_experiment(session, trash_dir / exp_id, exp_id)
+
+
+def _migrate_one_experiment(session: Session, exp_dir: Path, exp_id: str) -> None:
+    meta = safe_read_yaml(exp_dir, META_YAML)
+    if meta is None:
+        return
+
+    db_exp_id = int(exp_id)
+
+    session.add(
+        SqlExperiment(
+            experiment_id=db_exp_id,
+            name=meta.get("name", f"experiment_{exp_id}"),
+            artifact_location=meta.get("artifact_location"),
+            lifecycle_stage=meta.get("lifecycle_stage", "active"),
+            creation_time=meta.get("creation_time"),
+            last_update_time=meta.get("last_update_time"),
+        )
+    )
+    bump("experiments")
+
+    for key, value in read_tag_files(exp_dir / "tags").items():
+        session.add(
+            SqlExperimentTag(
+                key=key,
+                value=value,
+                experiment_id=db_exp_id,
+            )
+        )
+        bump("experiment_tags")
+
+
+# ── Phase 2: Runs + Params + Tags + Metrics + LatestMetrics ──────────────────
+
+RESERVED_FOLDERS = {"tags", "datasets", "traces", "models", ".trash"}
+
+
+def migrate_runs(session: Session, mlruns: Path) -> None:
+    for exp_id in list_experiment_ids(mlruns):
+        _migrate_runs_in_dir(session, mlruns / exp_id, int(exp_id))
+
+    trash_dir = mlruns / ".trash"
+    for exp_id in list_experiment_ids(trash_dir):
+        _migrate_runs_in_dir(session, trash_dir / exp_id, int(exp_id))
+
+
+def _migrate_runs_in_dir(session: Session, exp_dir: Path, exp_id: int) -> None:
+    for name in list_subdirs(exp_dir):
+        if name in RESERVED_FOLDERS:
+            continue
+        run_dir = exp_dir / name
+        if not (run_dir / META_YAML).is_file():
+            continue
+        _migrate_one_run(session, run_dir, exp_id)
+
+
+def _migrate_one_run(session: Session, run_dir: Path, exp_id: int) -> None:
+    meta = safe_read_yaml(run_dir, META_YAML)
+    if meta is None:
+        return
+
+    run_uuid = meta.get("run_uuid") or meta.get("run_id")
+    if not run_uuid:
+        return
+
+    status_raw = meta.get("status", RunStatus.RUNNING)
+    status = RunStatus.to_string(status_raw) if isinstance(status_raw, int) else str(status_raw)
+
+    session.add(
+        SqlRun(
+            run_uuid=run_uuid,
+            name=meta.get("run_name") or meta.get("name"),
+            source_type=(
+                meta.get("source_type", "LOCAL")
+                if isinstance(meta.get("source_type"), str)
+                else "LOCAL"
+            ),
+            source_name=meta.get("source_name", ""),
+            entry_point_name=meta.get("entry_point_name", ""),
+            user_id=meta.get("user_id"),
+            status=status,
+            start_time=meta.get("start_time"),
+            end_time=meta.get("end_time"),
+            deleted_time=meta.get("deleted_time"),
+            source_version=meta.get("source_version", ""),
+            lifecycle_stage=meta.get("lifecycle_stage", "active"),
+            artifact_uri=meta.get("artifact_uri"),
+            experiment_id=exp_id,
+        )
+    )
+    bump("runs")
+
+    # Params
+    for key, value in read_tag_files(run_dir / "params").items():
+        session.add(
+            SqlParam(
+                key=key,
+                value=value,
+                run_uuid=run_uuid,
+            )
+        )
+        bump("params")
+
+    # Tags
+    for key, value in read_tag_files(run_dir / "tags").items():
+        session.add(
+            SqlTag(
+                key=key,
+                value=value,
+                run_uuid=run_uuid,
+            )
+        )
+        bump("tags")
+
+    # Metrics + LatestMetrics
+    _migrate_run_metrics(session, run_dir / "metrics", run_uuid)
+
+
+def _sanitize_metric_value(val: float) -> tuple[bool, float]:
+    is_nan = math.isnan(val)
+    if is_nan:
+        return True, 0.0
+    if math.isinf(val):
+        return False, 1.7976931348623157e308 if val > 0 else -1.7976931348623157e308
+    return False, val
+
+
+def _parse_metric_line(metric_line: str) -> tuple[int, float, int]:
+    parts = metric_line.strip().split(" ")
+    ts = int(parts[0])
+    val = float(parts[1])
+    step = int(parts[2]) if len(parts) >= 3 else 0
+    return ts, val, step
+
+
+def _migrate_run_metrics(session: Session, metrics_dir: Path, run_uuid: str) -> None:
+    all_metrics = read_metric_lines(metrics_dir)
+
+    for key, lines in all_metrics.items():
+        # Track the "latest" metric for this key: max by (step, timestamp, value)
+        latest: tuple[int, int, float] | None = None  # (step, timestamp, value)
+        latest_is_nan = False
+
+        for line in lines:
+            ts, raw_val, step = _parse_metric_line(line)
+            is_nan, db_val = _sanitize_metric_value(raw_val)
+
+            session.add(
+                SqlMetric(
+                    key=key,
+                    value=db_val,
+                    timestamp=ts,
+                    step=step,
+                    is_nan=is_nan,
+                    run_uuid=run_uuid,
+                )
+            )
+            bump("metrics")
+
+            # For latest_metrics: NaN comparison uses 0 as proxy value
+            cmp_val = 0.0 if is_nan else db_val
+            cmp_tuple = (step, ts, cmp_val)
+            if latest is None or cmp_tuple > latest:
+                latest = cmp_tuple
+                latest_is_nan = is_nan
+
+        if latest is not None:
+            l_step, l_ts, l_val = latest
+            session.add(
+                SqlLatestMetric(
+                    key=key,
+                    value=0.0 if latest_is_nan else l_val,
+                    timestamp=l_ts,
+                    step=l_step,
+                    is_nan=latest_is_nan,
+                    run_uuid=run_uuid,
+                )
+            )
+            bump("latest_metrics")
+
+
+# ── Phase 3: Datasets + Inputs + InputTags ───────────────────────────────────
+
+
+def migrate_datasets(session: Session, mlruns: Path) -> None:
+    for exp_id in list_experiment_ids(mlruns):
+        _migrate_datasets_for_experiment(session, mlruns / exp_id, int(exp_id))
+
+    trash_dir = mlruns / ".trash"
+    for exp_id in list_experiment_ids(trash_dir):
+        _migrate_datasets_for_experiment(session, trash_dir / exp_id, int(exp_id))
+
+
+def _migrate_datasets_for_experiment(session: Session, exp_dir: Path, exp_id: int) -> None:
+    datasets_dir = exp_dir / "datasets"
+    if not datasets_dir.is_dir():
+        return
+
+    dataset_uuid_map: dict[str, str] = {}  # dataset_dir_name -> dataset_uuid
+
+    for ds_dir_name in list_subdirs(datasets_dir):
+        meta = safe_read_yaml(datasets_dir / ds_dir_name, META_YAML)
+        if meta is None:
+            continue
+
+        ds_uuid = meta.get("dataset_uuid") or str(uuid.uuid4())
+        dataset_uuid_map[ds_dir_name] = ds_uuid
+
+        session.add(
+            SqlDataset(
+                dataset_uuid=ds_uuid,
+                experiment_id=exp_id,
+                name=meta.get("name", ""),
+                digest=meta.get("digest", ""),
+                dataset_source_type=meta.get("source_type", ""),
+                dataset_source=meta.get("source", ""),
+                dataset_schema=meta.get("schema"),
+                dataset_profile=meta.get("profile"),
+            )
+        )
+        bump("datasets")
+
+    # Scan runs in this experiment for inputs
+    for run_name in list_subdirs(exp_dir):
+        if run_name in RESERVED_FOLDERS:
+            continue
+        inputs_dir = exp_dir / run_name / "inputs"
+        if not inputs_dir.is_dir():
+            continue
+
+        for input_dir_name in list_subdirs(inputs_dir):
+            input_meta = safe_read_yaml(inputs_dir / input_dir_name, META_YAML)
+            if input_meta is None:
+                continue
+
+            source_type = input_meta.get("source_type", "DATASET")
+            source_id = input_meta.get("source_id", "")
+
+            ds_uuid = dataset_uuid_map.get(source_id)
+            if source_type == "DATASET" and ds_uuid is None:
+                continue
+
+            input_uuid = str(uuid.uuid4())
+            session.add(
+                SqlInput(
+                    input_uuid=input_uuid,
+                    source_type=source_type,
+                    source_id=ds_uuid if source_type == "DATASET" else source_id,
+                    destination_type="RUN",
+                    destination_id=run_name,
+                )
+            )
+            bump("inputs")
+
+            input_tags = input_meta.get("tags", {})
+            for tag_name, tag_value in input_tags.items():
+                session.add(
+                    SqlInputTag(
+                        input_uuid=input_uuid,
+                        name=tag_name,
+                        value=str(tag_value),
+                    )
+                )
+                bump("input_tags")
+
+
+# ── Phase 4: Traces + TraceTags + TraceMetadata ──────────────────────────────
+
+
+def migrate_traces(session: Session, mlruns: Path) -> None:
+    for exp_id in list_experiment_ids(mlruns):
+        _migrate_traces_for_experiment(session, mlruns / exp_id, int(exp_id))
+
+    trash_dir = mlruns / ".trash"
+    for exp_id in list_experiment_ids(trash_dir):
+        _migrate_traces_for_experiment(session, trash_dir / exp_id, int(exp_id))
+
+
+def _migrate_traces_for_experiment(session: Session, exp_dir: Path, exp_id: int) -> None:
+    traces_dir = exp_dir / "traces"
+    if not traces_dir.is_dir():
+        return
+
+    for trace_dir_name in list_subdirs(traces_dir):
+        trace_dir = traces_dir / trace_dir_name
+        if not (trace_dir / "trace_info.yaml").is_file():
+            continue
+
+        meta = safe_read_yaml(trace_dir, "trace_info.yaml")
+        if meta is None:
+            continue
+
+        # V2 uses request_id, V3 uses trace_id
+        trace_id = meta.get("trace_id") or meta.get("request_id") or trace_dir_name
+
+        # V2 uses timestamp_ms, V3 uses request_time (proto timestamp string)
+        timestamp_ms = meta.get("timestamp_ms")
+        if timestamp_ms is None:
+            request_time = meta.get("request_time")
+            if isinstance(request_time, int):
+                timestamp_ms = request_time
+            elif isinstance(request_time, str):
+                try:
+                    from google.protobuf.timestamp_pb2 import Timestamp
+
+                    ts = Timestamp()
+                    ts.FromJsonString(request_time)
+                    timestamp_ms = ts.ToMilliseconds()
+                except Exception:
+                    timestamp_ms = 0
+            else:
+                timestamp_ms = 0
+
+        # V2 uses execution_time_ms, V3 uses execution_duration_ms
+        execution_time_ms = meta.get("execution_time_ms") or meta.get("execution_duration_ms")
+
+        # Status: V2 has status as string like "OK", V3 has state
+        status = meta.get("status") or meta.get("state", "OK")
+
+        session.add(
+            SqlTraceInfo(
+                request_id=trace_id,
+                experiment_id=exp_id,
+                timestamp_ms=timestamp_ms,
+                execution_time_ms=execution_time_ms,
+                status=status,
+                client_request_id=meta.get("client_request_id"),
+                request_preview=meta.get("request_preview"),
+                response_preview=meta.get("response_preview"),
+            )
+        )
+        bump("traces")
+
+        # Trace tags
+        for key, value in read_tag_files(trace_dir / "tags").items():
+            session.add(
+                SqlTraceTag(
+                    key=key,
+                    value=value,
+                    request_id=trace_id,
+                )
+            )
+            bump("trace_tags")
+
+        # Trace request metadata
+        for key, value in read_tag_files(trace_dir / "request_metadata").items():
+            session.add(
+                SqlTraceMetadata(
+                    key=key,
+                    value=value,
+                    request_id=trace_id,
+                )
+            )
+            bump("trace_metadata")
+
+
+# ── Phase 5: Assessments ─────────────────────────────────────────────────────
+
+
+def migrate_assessments(session: Session, mlruns: Path) -> None:
+    for exp_id in list_experiment_ids(mlruns):
+        _migrate_assessments_for_experiment(session, mlruns / exp_id)
+
+    trash_dir = mlruns / ".trash"
+    for exp_id in list_experiment_ids(trash_dir):
+        _migrate_assessments_for_experiment(session, trash_dir / exp_id)
+
+
+def _migrate_assessments_for_experiment(session: Session, exp_dir: Path) -> None:
+    traces_dir = exp_dir / "traces"
+    if not traces_dir.is_dir():
+        return
+
+    for trace_dir_name in list_subdirs(traces_dir):
+        trace_dir = traces_dir / trace_dir_name
+        assessments_dir = trace_dir / "assessments"
+        if not assessments_dir.is_dir():
+            continue
+
+        trace_meta = safe_read_yaml(trace_dir, "trace_info.yaml")
+        if trace_meta is None:
+            continue
+        trace_id = trace_meta.get("trace_id") or trace_meta.get("request_id") or trace_dir_name
+
+        for filename in list_files(assessments_dir):
+            if not filename.endswith(".yaml"):
+                continue
+            assessment_id = filename[:-5]  # strip .yaml
+            meta = safe_read_yaml(assessments_dir, filename)
+            if meta is None:
+                continue
+
+            _migrate_one_assessment(session, meta, trace_id, assessment_id)
+
+
+def _migrate_one_assessment(
+    session: Session, meta: dict[str, object], trace_id: str, assessment_id: str
+) -> None:
+    feedback_data = meta.get("feedback")
+    expectation_data = meta.get("expectation")
+
+    if feedback_data is not None:
+        assessment_type = "feedback"
+        value_json = json.dumps(feedback_data.get("value"))
+        error_data = feedback_data.get("error")
+        error_json = json.dumps(error_data) if error_data else None
+    elif expectation_data is not None:
+        assessment_type = "expectation"
+        value_json = json.dumps(expectation_data.get("value"))
+        error_json = None
+    else:
+        return
+
+    source = meta.get("source", {})
+    source_type = source.get("source_type", "CODE")
+    source_id = source.get("source_id")
+
+    create_time = meta.get("create_time_ms", 0)
+    last_update_time = meta.get("last_update_time_ms", create_time)
+
+    assessment_metadata = meta.get("metadata")
+    metadata_json = json.dumps(assessment_metadata) if assessment_metadata else None
+
+    session.add(
+        SqlAssessments(
+            assessment_id=meta.get("assessment_id") or assessment_id,
+            trace_id=trace_id,
+            name=meta.get("assessment_name", meta.get("name", "")),
+            assessment_type=assessment_type,
+            value=value_json,
+            error=error_json,
+            created_timestamp=create_time,
+            last_updated_timestamp=last_update_time,
+            source_type=source_type,
+            source_id=source_id,
+            run_id=meta.get("run_id"),
+            span_id=meta.get("span_id"),
+            rationale=meta.get("rationale"),
+            overrides=meta.get("overrides"),
+            valid=meta.get("valid", True),
+            assessment_metadata=metadata_json,
+        )
+    )
+    bump("assessments")
+
+
+# ── Phase 6: Logged Models + Params + Tags + Metrics ────────────────────────
+
+
+def migrate_logged_models(session: Session, mlruns: Path) -> None:
+    for exp_id in list_experiment_ids(mlruns):
+        _migrate_logged_models_for_experiment(session, mlruns / exp_id, int(exp_id))
+
+    trash_dir = mlruns / ".trash"
+    for exp_id in list_experiment_ids(trash_dir):
+        _migrate_logged_models_for_experiment(session, trash_dir / exp_id, int(exp_id))
+
+
+def _migrate_logged_models_for_experiment(session: Session, exp_dir: Path, exp_id: int) -> None:
+    models_dir = exp_dir / "models"
+    if not models_dir.is_dir():
+        return
+
+    for model_dir_name in list_subdirs(models_dir):
+        model_dir = models_dir / model_dir_name
+        meta = safe_read_yaml(model_dir, META_YAML)
+        if meta is None:
+            continue
+
+        model_id = meta.get("model_id", model_dir_name)
+
+        # Status may be stored as an integer enum or string
+        status_raw = meta.get("status", 1)  # 1 = PENDING typically
+        if isinstance(status_raw, str):
+            from mlflow.entities.logged_model_status import LoggedModelStatus
+
+            try:
+                status_raw = LoggedModelStatus[status_raw].value
+            except (KeyError, AttributeError):
+                status_raw = 1
+        status = int(status_raw)
+
+        session.add(
+            SqlLoggedModel(
+                model_id=model_id,
+                experiment_id=exp_id,
+                name=meta.get("name", ""),
+                artifact_location=meta.get("artifact_location", ""),
+                creation_timestamp_ms=meta.get("creation_timestamp", 0),
+                last_updated_timestamp_ms=meta.get("last_updated_timestamp", 0),
+                status=status,
+                lifecycle_stage=meta.get("lifecycle_stage", "active"),
+                model_type=meta.get("model_type"),
+                source_run_id=meta.get("source_run_id"),
+                status_message=meta.get("status_message"),
+            )
+        )
+        bump("logged_models")
+
+        # Logged model params
+        for key, value in read_tag_files(model_dir / "params").items():
+            session.add(
+                SqlLoggedModelParam(
+                    model_id=model_id,
+                    experiment_id=exp_id,
+                    param_key=key,
+                    param_value=value,
+                )
+            )
+            bump("logged_model_params")
+
+        # Logged model tags
+        for key, value in read_tag_files(model_dir / "tags").items():
+            session.add(
+                SqlLoggedModelTag(
+                    model_id=model_id,
+                    experiment_id=exp_id,
+                    tag_key=key,
+                    tag_value=value,
+                )
+            )
+            bump("logged_model_tags")
+
+        # Logged model metrics
+        _migrate_logged_model_metrics(session, model_dir / "metrics", model_id, exp_id)
+
+
+def _migrate_logged_model_metrics(
+    session: Session, metrics_dir: Path, model_id: str, exp_id: int
+) -> None:
+    all_metrics = read_metric_lines(metrics_dir)
+
+    for key, lines in all_metrics.items():
+        for line in lines:
+            parts = line.strip().split(" ")
+            # Logged model metric format: timestamp value step run_id [dataset_name dataset_digest]
+            if len(parts) not in (4, 6):
+                continue
+
+            ts = int(parts[0])
+            val = float(parts[1])
+            step = int(parts[2])
+            run_id = parts[3]
+            dataset_name = parts[4] if len(parts) == 6 else None
+            dataset_digest = parts[5] if len(parts) == 6 else None
+
+            session.add(
+                SqlLoggedModelMetric(
+                    model_id=model_id,
+                    metric_name=key,
+                    metric_timestamp_ms=ts,
+                    metric_step=step,
+                    metric_value=val,
+                    experiment_id=exp_id,
+                    run_id=run_id,
+                    dataset_uuid=None,
+                    dataset_name=dataset_name,
+                    dataset_digest=dataset_digest,
+                )
+            )
+            bump("logged_model_metrics")
